Simplify demo app by getting rid of the server. Use token factory to generate tokens instead
