+++ package.json
@@ -4,7 +4,7 @@
   "description": "Sample App for JavaScript SDK view layer for Onfido identity verification",
   "scripts": {
     "prestart": "npm run build",
-    "start": "babel-node server/server.js",
+    "start": "webpack-dev-server --inline --hot --progress --https",
     "build": "NODE_ENV=production webpack -p --progress",
     "lint": "eslint {src,test}",
     "link:sdk": "npm link ../onfido-sdk-ui",
@@ -43,10 +43,7 @@
   },
   "dependencies": {
     "@babel/runtime-corejs3": "^7.4.4",
-    "express": "^4.14.0",
     "onfido-sdk-ui": "^5.9.0",
-    "pem": "^1.14.4",
-    "unirest": "^0.6.0",
-    "yarn": "^1.22.4"
+    "webpack-dev-server": "^3.10.3"
   }
 }
